+++ system/docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Version 3.0.4 (2013-XX-XX)
 --------------------------
 
 ### Fixed
+Correctly determin the last run of the command scheduler (see #5278).
+
+### Fixed
 Make the jQuery accordion behave like the MooTools version (see #5251).
 
 ### Fixed