+++ system/docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Version 3.0.4 (2013-XX-XX)
 --------------------------
 
 ### Fixed
+Removed the "spaceToUnderscore" option from all alias fields (see #5266).
+
+### Fixed
 The media content element now supports .ogg files (see #5282).
 
 ### Fixed